requirements.txt
git+https://github.com/python-ivi/python-usbtmc.git@master
pyserial>=3.5
pyusb>=1.2.1
